requirements.txt
fastapi
uvicorn[standard]
gitpython
google-generativeai
pydantic
python-multipart
python-dotenv